requirements.txt
monai
torchvision
SimpleITK
requests
libtorrent
scikit-learn
numpy==1.26.4
medpy